Restructure my files system to have each component and it's corresponding css file in the same folder
